added tests for classification metrics
